test: 로그인 스텝 클래스로 변경
